// packages/core/src/Table/hooks/useTable.d.ts
type Accessor<D extends object> = (
  originalRow: D,
  index: number,
  sub: {
    subRows: D[];
    depth: number;
    data: D[];
  }
) => CellValue;

type StringKey<D> = Extract<keyof D, string>;
type IdType<D> = StringKey<D> | string;
type CellValue<V = any> = V;

interface HvColumnInterfaceBasedOnValue<D extends object = Record<string, unknown>, V = any> {
  Cell?: Renderer<HvCellProps<D, V>>;
}

interface HvColumnGroupInterface<D extends object> {
  columns: Array<HvTableColumnConfig<D>>;
}

type HvColumnGroup<D extends object = Record<string, unknown>> = HvTableColumnOptions<D> &
  HvColumnGroupInterface<D> &
  (
    | { Header: string }
    | ({ id: IdType<D> } & {
        Header: Renderer<HeaderProps<D>>;
      })
  ) & { accessor?: Accessor<D> }; // Not used, but needed for backwards compatibility

type ValueOf<T> = T[keyof T];

// The accessors like `foo.bar` are not supported, use functions instead
type HvColumnWithStrictAccessor<D extends object = Record<string, unknown>> =
  HvTableColumnOptions<D> &
    ValueOf<{
      [K in keyof D]: {
        accessor: K;
      } & HvColumnInterfaceBasedOnValue<D, D[K]>;
    }>;

type HvColumnWithLooseAccessor<D extends object = Record<string, unknown>> =
  HvTableColumnOptions<D> &
    HvColumnInterfaceBasedOnValue<D> &
    (
      | { Header: string }
      | { id: IdType<D> }
      | { accessor: keyof D extends never ? IdType<D> : never }
    ) & { accessor?: keyof D extends never ? IdType<D> | Accessor<D> : Accessor<D> };

export type HvCellProps<D extends object, V = any> = CellProps<D, V> & {
  column: HvColumnInstance<D>;
  row: HvRowInstance<D>;
  cell: HvCellInstance<D, V>;
};

export type HvTableDefinitionConfig<D extends object = Record<string, unknown>> = HvTableOptions<D>;

export type HvTableColumnConfig<D extends object = Record<string, unknown>> =
  | HvColumnGroup<D>
  | HvColumnWithLooseAccessor<D>
  | HvColumnWithStrictAccessor<D>;
// ...
export interface HvTableColumnOptions<D extends object = Record<string, unknown>>
  extends ColumnInterface<D>,
    UseFiltersColumnOptions<D>,
    UseGroupByColumnOptions<D>,
    UseResizeColumnsColumnOptions<D>,
    UseSortByColumnOptions<D>,
    UseHvTableStylesColumnOptions {}
// ...
export interface HvColumnInstance<D extends object = Record<string, unknown>>
  extends ColumnInstance<D>,
    Omit<HvTableColumnOptions<D>, "id">,
    UseFiltersColumnProps<D>,
    UseGroupByColumnProps<D>,
    UseResizeColumnsColumnProps<D>,
    UseSortByColumnProps<D> {
  Cell?: Renderer<HvCellProps<D>>;
  columns: Array<HvColumnInstance<D>>;
  parent: HvColumnInstance<D>; // not documented
  placeholderOf?: HvColumnInstance;
  getHeaderProps: (propGetter?: HeaderPropGetter<D>) => HvUseTableHeaderProps;
  getFooterProps: (propGetter?: FooterPropGetter<D>) => HvUseTableFooterProps;
}

export interface HvRowInstance<D extends object = Record<string, unknown>>
  extends Row<D>,
    UseGroupByRowProps<D>,
    UseHvRowSelectionRowInstance,
    UseHvRowExpandRowInstance<D> {
  cells: Array<HvCellInstance<D>>;
  allCells: Array<HvCellInstance<D>>;
  getRowProps: (propGetter?: RowPropGetter<D>) => HvUseTableRowProps;
  index: number;
  original: D;
  id: string;
  subRows: Array<HvRowInstance<D>>;
}

export interface HvCellInstance<D extends object = Record<string, unknown>, V = any>
  extends Cell<D, V>,
    UseGroupByCellProps<D> {
  column: HvColumnInstance<D>;
  row: HvRowInstance<D>;
  getCellProps: (propGetter?: CellPropGetter<D>) => HvUseTableCellProps;
}
// #endregion

// #region PROPS
export interface HvUseTableProps extends TableProps, UseHvTableStickyTableProps {}

export interface HvUseTableHeaderProps
  extends TableHeaderProps,
    UseHvTableStylesTableCellProps,
    UseHvTableStickyColumnProps,
    UseHvHeaderGroupsColumnProps,
    UseHvResizeColumnProps,
    UseHvSortByColumnProps {}
export interface HvUseTableFooterProps extends TableFooterProps, UseHvTableStylesTableCellProps {}

export interface HvUseTableRowProps
  extends TableRowProps,
    UseHvTableStylesTableRowProps,
    UseHvRowSelectionTableRowProps,
    UseHvRowExpandTableRowProps {}

export interface HvUseTableCellProps
  extends TableCellProps,
    UseHvTableStylesTableCellProps,
    UseHvTableStickyCellProps,
    UseHvHeaderGroupsCellProps,
    UseHvResizeTableCellProps,
    UseHvSortByTableCellProps {}